BREAKING: Rebuild files without change to line height
Changes upstream affected our override of the 'leading' variable to the
effect that some point sizes would crop accents over capital letters.

Also, the "wide" variants no longer occupied the same lines as the
others, which is a regression based on what we had before.  I do not
want the "wide" ones to be a different font.

This is not yet part of a release.  I might make further changes.