Revise the "motion" variants
Removes some of the curly or odd characters.  I think this fixes the
style for good.  It makes the "motion" variants look closer to the
standard iosevka-comfy style.

As with commit 37e6ced, this is not yet part of a new release.  I might
still make further changes.  I think the 'f' may need to be changed in
these "motion" variants.  It feels too fancy right now, but I am not
sure yet.