Rebuild "motion" fonts per ad36dc9, 5d6b16d, 135fcf2
As with commits ccc9c4c, 37e6ced, this does not yet form part of a
release.  I am testing all the fonts to ensure everything works as
intended.  I will prepare a release when the time is right.

For the underlying reason, refer to the commit message of 37e6ced.