Update to Iosevka Comfy version 2.1.0
This is a small release that fixes a detail in the `m` character of
all the "motion" variants. It now has a tail at the bottom right
corner in addition to the top left serif. This is the same design used
wherever relevant, such as in the `n` and `u` glyphs.

Other than that, the rebuild is an attempt to fix the broken files
that appear on NixOS, as reported by Damien Cassou in issue 17:
<https://github.com/protesilaos/iosevka-comfy/issues/17>. I do not
know if this fixes the problem, but I cannot think of a better
approach. This is not reproducible on my end (Debian stable).